Remove TypeTag from Column; move to UntypedColumn.
This adds a new abstraction for dealing with un-typed columns, rather
than trying to force Column[_] to do both. The new UntypedColumn has
some basic operations that don't rely on the type and a new cast method
that can be used to extract a typed Column from an untyped one.