Consolidate selection of columns from frame.
Duplicating methods, 1 for arity 1 and 1 for arity N was annoying. Also,
for a type that should be able to handle both row- and column-oriented
operations, I wanted a way to disambiguate things like `map` without
resorting to ugly naming conventions.

This moves the existing column operations into a new ColumnSelector
type. The frame then just provides a way to obtain a default
ColumnSelector.

Finally, we've now parameterized the row extractors on a Size type. This
let's us handle both the single arity case and N-arity case uniformly.
The column selector carries this as well, which means some methods are
available only in certain contexts, which is nice.