Add Series zipMap benchmark.
This adds a zipMap benchmark for series that breaks down where time is
spent. The short of it is that cogrouping takes the most time, not
surprisingly. It's not specialized and it is complicated. However, I'm
not sure if it is worth specializing. String, LocalDates, etc, these
things seem more plausible as keys than Double or Int.